Move TokenBucket implementation to token_bucket.rs
The `lib.rs` file is normally used to expose library's public interface
to external consumers. We probably don't want to maintain the algorithm
implementation in it because there might be more algorithms implemented
later on. This patch essentially moves `TokenBucket` struct and its
implementation out from `lib.rs` to `token_bucket.rs`.
